refactor: rename task service package and update docker
